+++ rtc-sentinel/client/eslint.config.js
@@ -4,7 +4,7 @@ import reactRefresh from 'eslint-plugin-react-refresh';
 import tseslint from 'typescript-eslint';
 
 export default tseslint.config(
-  { ignores: ['dist', 'coverage'] },
+  { ignores: ['dist', 'coverage', '*.cjs'] },
   js.configs.recommended,
   ...tseslint.configs.recommended,
   {
@@ -16,4 +16,3 @@ export default tseslint.config(
     },
   },
 );
-